requirements.txt
alabaster==0.7.13
Babel==2.12.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dlib==19.24.2
dnspython==2.3.0
docutils==0.20.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
Pygments==2.15.1
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==2.0.3
Werkzeug==2.3.4
zipp==3.15.0
